api/requirements.txt
asgiref==3.3.1
Django==3.1.3
django-cors-headers==3.5.0
django-leaflet==0.27.1
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-gis==0.16
GDAL==3.1.4
leaflet==0.0.3
psycopg2==2.8.6
pytz==2020.4
sqlparse==0.4.1
